requirements.txt
selenium~=4.16.0
PyYAML~=6.0